Run each integration test in its own process
